Added Functionality to group sub-pixels into pixels
Created a new class called MultiSpanIterator, that iterates over an Iterable of TimeSeriesSpan.

Created a new class called PixelAggregator, that takes sub-pixel intervals from the MultiSpanIterator class, and groups them by pixel. It then returns the min/max value of that pixel.

Need to implement functionality to take into account the overlapping intervals at the boundaries of pixel. As it stands now, they are grouped in the first pixel they are founc on.